pratica37.1 - refatora TaskContext dividindo responsabilidades em múltiplos arquivos
