+++ praticas/pomodoro/src/components/CountDown/index.tsx
@@ -1,5 +1,5 @@
 import styles from './styles.module.css';
-import { useTaskContext } from '../../contexts/TaskContext';
+import { useTaskContext } from '../../contexts/TaskContext/useTaskContext';
 
 export function CountDown() {
   // Puxamos apenas o 'state' de dentro do nosso contexto
