--- praticas/pomodoro/src/contexts/TaskContext.tsx
@@ -1,49 +1,0 @@
-import { createContext, useContext, useState } from 'react';
-import type { TaskStateModel } from '../../src/models/TaskStateModel';
-
-const initialState: TaskStateModel = {
-  tasks: [],
-  secondsRemaining: 0,
-  formattedSecondsRemaining: '00:00',
-  activeTask: null,
-  currentCycle: 0,
-  config: {
-    workTime: 25,
-    shortBreakTime: 5,
-    longBreakTime: 15,
-  },
-};
-
-type TaskContextProps = {
-  state: TaskStateModel;
-  setState: React.Dispatch<React.SetStateAction<TaskStateModel>>;
-};
-
-// Valor inicial de fallback (caso alguém use o contexto fora do provider)
-const initialContextValue = {
-  state: initialState,
-  setState: () => {},
-};
-
-export const TaskContext = createContext<TaskContextProps>(initialContextValue);
-
-type TaskContextProviderProps = {
-  children: React.ReactNode;
-};
-
-export function TaskContextProvider({ children }: TaskContextProviderProps) {
-  // O Estado agora mora DENTRO do Provider!
-  const [state, setState] = useState(initialState);
-
-  return (
-    // Passamos tanto o state quanto a função de atualizar (setState) para todos os filhos
-    <TaskContext.Provider value={{ state, setState }}>
-      {children}
-    </TaskContext.Provider>
-  );
-}
-
-// Hook customizado para facilitar o uso
-export function useTaskContext() {
-  return useContext(TaskContext);
-}
