+++ praticas/pomodoro/src/contexts/TaskContext/TaskContext.ts
@@ -1,0 +1,15 @@
+import { createContext } from 'react';
+import type { TaskStateModel } from '../../models/TaskStateModel';
+import { initialTaskState } from './initialTaskState';
+
+type TaskContextProps = {
+  state: TaskStateModel;
+  setState: React.Dispatch<React.SetStateAction<TaskStateModel>>;
+};
+
+const initialContextValue = {
+  state: initialTaskState,
+  setState: () => {},
+};
+
+export const TaskContext = createContext<TaskContextProps>(initialContextValue);
