+++ praticas/pomodoro/src/contexts/TaskContext/TaskContextProvider.tsx
@@ -1,0 +1,17 @@
+import { useState } from 'react';
+import { initialTaskState } from './initialTaskState';
+import { TaskContext } from './TaskContext';
+
+type TaskContextProviderProps = {
+  children: React.ReactNode;
+};
+
+export function TaskContextProvider({ children }: TaskContextProviderProps) {
+  const [state, setState] = useState(initialTaskState);
+
+  return (
+    <TaskContext.Provider value={{ state, setState }}>
+      {children}
+    </TaskContext.Provider>
+  );
+}
